+++ CHANGELOG.md
@@ -19,6 +19,7 @@ For a full diff see [`2.2.0...3.0.0`][2.2.0...3.0.0].
 - Removed `Exception` suffix from all exceptions ([#668]), by [@dependabot]
 - Renamed `NormalizerInterface` to `Normalizer` ([#669]), by [@dependabot]
 - Renamed `Format\Formatter` to `Format\DefaultFormatter` ([#672]), by [@dependabot]
+- Renamed `Format\FormatterInterface` to `Format\Formatter` ([#673]), by [@dependabot]
 
 ## [`2.2.0`][2.2.0]
 
@@ -495,6 +496,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#668]: https://github.com/ergebnis/json-normalizer/pull/668
 [#669]: https://github.com/ergebnis/json-normalizer/pull/669
 [#672]: https://github.com/ergebnis/json-normalizer/pull/672
+[#673]: https://github.com/ergebnis/json-normalizer/pull/673
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@dependabot]: https://github.com/dependabot